tostring fix for uuid
git-svn-id: https://realxtend-naali.googlecode.com/svn/trunk@290 5b2332b8-efa3-11de-8684-7d64432d61a3

+++ RexCommon/RexUUID.cpp
@@ -76,19 +76,19 @@ namespace RexTypes
         stringstream str;
         int i = 0;
     
-        for(int j = 0; j < 4; ++j) str << hex << data[i++];
+        for(int j = 0; j < 4; ++j) str << hex << (int)data[i++];
         str << "-";
     
-        for(int j = 0; j < 2; ++j) str << hex << data[i++];
+        for(int j = 0; j < 2; ++j) str << hex << (int)data[i++];
         str << "-";
     
-        for(int j = 0; j < 2; ++j) str << hex << data[i++];
+        for(int j = 0; j < 2; ++j) str << hex << (int)data[i++];
         str << "-";
     
-        for(int j = 0; j < 2; ++j) str << hex << data[i++];
+        for(int j = 0; j < 2; ++j) str << hex << (int)data[i++];
         str << "-";
     
-        for(int j = 0; j < 6; ++j) str << hex << data[i++];
+        for(int j = 0; j < 6; ++j) str << hex << (int)data[i++];
     
         return str.str();
     }
